Ajout des requetes dans les différents manager et correciton des erreurs
Reprise des fonctions contenant les requetes pour chaque éléments puis modification des liens dans les code
